+++ packages/frontend/src/components/DashboardTabs/gridUtils.ts
@@ -11,6 +11,10 @@ export type ResponsiveGridLayoutProps = {
 };
 
 const DEFAULT_COLS = 36;
+/**
+ * Row height: fontSize * lineHeight + padding + borders
+ */
+const DEFAULT_ROW_HEIGHT = 14 * 1.5 + 16 * 2 + 2;
 
 export const getReactGridLayoutConfig = (
     tile: DashboardTile,
@@ -55,5 +59,5 @@ export const getResponsiveGridLayoutProps = ({
         md: 30,
         sm: stackVerticallyOnSmallestBreakpoint ? 1 : 18,
     },
-    rowHeight: 50,
+    rowHeight: DEFAULT_ROW_HEIGHT,
 });