+++ packages/frontend/vite-env.d.ts
@@ -1,4 +1,6 @@
 /// <reference types="vite/client" />
 /// <reference types="vite-plugin-svgr/client" />
 
-declare const __APP_VERSION__: string
+declare const __APP_VERSION__: string;
+declare const REACT_SCAN_ENABLED: boolean;
+declare const REACT_QUERY_DEVTOOLS_ENABLED: boolean;